LP-2523 move non-generic fixtures to applications app
