+++ openedx/adg/lms/applications/tests/conftest.py
@@ -1,7 +1,7 @@
 """
 File containing common fixtures used across different test modules
 """
-from datetime import date
+from datetime import date, timedelta
 
 import pytest
 
@@ -14,6 +14,8 @@ from openedx.adg.lms.applications.admin import (
 )
 from openedx.adg.lms.applications.constants import ADG_ADMIN_GROUP_NAME
 from openedx.adg.lms.applications.models import UserApplication
+from openedx.adg.lms.applications.tests.factories import MultilingualCourseGroupFactory
+from openedx.core.djangoapps.content.course_overviews.tests.factories import CourseOverviewFactory
 
 from .constants import ADMIN_TYPE_SUPER_ADMIN, TITLE_BUSINESS_LINE_1, TITLE_BUSINESS_LINE_2
 from .factories import EducationFactory, UserApplicationFactory, WorkExperienceFactory
@@ -75,3 +77,32 @@ def admin_user(admin_type):
         return UserFactory(is_superuser=True)
     else:   # ADG admin
         return UserFactory(is_staff=True, groups=[GroupFactory(name=ADG_ADMIN_GROUP_NAME)])
+
+
+@pytest.fixture(name='courses')
+def course_overviews(current_time):
+    """
+    Fixture which return multiple courses
+    """
+    course_start_end_date = {
+        'start_date': current_time - timedelta(days=1),
+        'end_date': current_time + timedelta(days=1),
+    }
+    return {
+        'test_course1': CourseOverviewFactory(language='en', **course_start_end_date),
+        'test_course2': CourseOverviewFactory(language='ar', **course_start_end_date),
+    }
+
+
+@pytest.fixture(name='expired_course')
+def expired_course_overview(current_time):
+    return CourseOverviewFactory(
+        language='en',
+        start_date=current_time - timedelta(days=2),
+        end_date=current_time - timedelta(days=1),
+    )
+
+
+@pytest.fixture(name='course_group')
+def multilingual_course_group():
+    return MultilingualCourseGroupFactory()
